# restore performance improvements (part2)