@dimagimon 2015-01-12
There are 8 errors:

* corehq/ex-submodules/casexml/apps/phone/caselogic.py, line 400 - E501 line too long (116 > 115 characters)
* corehq/ex-submodules/casexml/apps/phone/caselogic.py, line 402 - E501 line too long (138 > 115 characters)
* corehq/ex-submodules/casexml/apps/phone/caselogic.py, line 421 - E501 line too long (118 > 115 characters)
* corehq/ex-submodules/casexml/apps/phone/tests/__init__.py, line 7 - F401 'BatchRestoreTests' imported but unused
* corehq/ex-submodules/casexml/apps/phone/tests/__init__.py, line 7 - F401 'MultiUserSyncTestBatched' imported but unused
* corehq/ex-submodules/casexml/apps/phone/tests/__init__.py, line 7 - F401 'OtaRestoreTestBatched' imported but unused
* corehq/ex-submodules/casexml/apps/phone/tests/__init__.py, line 7 - F401 'SyncTokenUpdateTestBatched' imported but unused
* corehq/ex-submodules/casexml/apps/phone/tests/test_sync_mode.py, line 44 - W291 trailing whitespace