@snopoke 2015-01-12
@czue I re-looked at this and the scenario you described won't be a problem since I always load all the case IDs from couch so any new ones still get included.

Added a unit test to verify.